StateTypeを使うように変更
