+++ database/session_test.go
@@ -191,7 +191,7 @@ func TestSessionRepository_Update(t *testing.T) {
 				Name:        "existing_session_name",
 				CreatorID:   "existing_user",
 				QueueHead:   1,
-				StateType:   "PLAY",
+				StateType:   entity.Play,
 				QueueTracks: []*entity.QueueTrack{},
 			},
 			wantErr: false,
@@ -203,7 +203,7 @@ func TestSessionRepository_Update(t *testing.T) {
 				Name:        "not_found_id_names",
 				CreatorID:   "existing_user",
 				QueueHead:   1,
-				StateType:   "PLAY",
+				StateType:   entity.Play,
 				QueueTracks: []*entity.QueueTrack{},
 			},
 			wantErr: true,
