+++ domain/entity/session.go
@@ -8,15 +8,40 @@ type Session struct {
 	Name        string
 	CreatorID   string
 	QueueHead   int
-	StateType   string
+	StateType   StateType
 	QueueTracks []*QueueTrack
 }
 
 // MoveToPlay はセッションのStateTypeをPlayに状態遷移します。ｓ
 func (s *Session) MoveToPlay() error {
-	if s.StateType == Pause.String() || s.StateType == Stop.String() {
-		s.StateType = Play.String()
+	if s.StateType == Pause || s.StateType == Stop {
+		s.StateType = Play
 		return nil
 	}
 	return fmt.Errorf("state type from %s to Play: %w", s.StateType, ErrChangeSessionStateNotPermit)
+}
+
+type StateType string
+
+const (
+	Play  StateType = "PLAY"
+	Pause StateType = "PAUSE"
+	Stop  StateType = "STOP"
+)
+
+var stateTypes = []StateType{Play, Pause, Stop}
+
+// NewStateType はstringから対応するStateTypeを生成します。
+func NewStateType(stateType string) (StateType, error) {
+	for _, st := range stateTypes {
+		if st.String() == stateType {
+			return st, nil
+		}
+	}
+	return "", fmt.Errorf("stateType = %s:%w", stateType, ErrInvalidStateType)
+}
+
+// String はfmt.Stringerを満たすメソッドです。
+func (st StateType) String() string {
+	return string(st)
 }
