+++ domain/entity/session_test.go
