--- domain/entity/state.go
@@ -1,28 +1,0 @@
-package entity
-
-import "fmt"
-
-type StateType string
-
-const (
-	Play  StateType = "PLAY"
-	Pause StateType = "PAUSE"
-	Stop  StateType = "STOP"
-)
-
-var stateTypes = []StateType{Play, Pause, Stop}
-
-// NewStateType はstringから対応するStateTypeを生成します。
-func NewStateType(stateType string) (StateType, error) {
-	for _, st := range stateTypes {
-		if st.String() == stateType {
-			return st, nil
-		}
-	}
-	return "", fmt.Errorf("stateType = %s:%w", stateType, ErrInvalidStateType)
-}
-
-// String はfmt.Stringerを満たすメソッドです。
-func (st StateType) String() string {
-	return string(st)
-}
